Add the httptrace feature
In order to be able to trace the latency of the connection. I use the httptrace package from golang.
If trace is nil, we can set a default one, that trace the basics (you can add more callback functions if you want)

On zipkin the big overview will give you some dots (with hover labels) :

![zipkin span view](https://drive.google.com/open?id=121PmPrZB1WuZjnMtVbeOps0ReEEr-rMw)

And if you open the span, you will see each log, and the time when it occured (allowing you to see the latency) :

![zipkin span log](https://drive.google.com/open?id=1Hfi1ewxbAR8iP-D13TypONdB-_xXzmBq)